+++ client/src/app/globals.css
@@ -2,7 +2,11 @@
 @tailwind components;
 @tailwind utilities;
 
+@layer base {
+
+
 :root {
-  --light: #333;
-  --light-red: #989766
+  --test: #333
+}
+
 }
