+++ client/src/app/page.js
@@ -1,0 +1,9 @@
+
+
+export default function Home() {
+  return (
+  <div>
+    <h1 className="bg-test">Hello</h1>
+  </div>
+  );
+}
