--- client/src/app/page.tsx
@@ -1,7 +1,0 @@
-import Image from "next/image";
-
-export default function Home() {
-  return (
-  <div className="bg-red">App</div>
-  );
-}
