--- client/tailwind.config.ts
@@ -1,19 +1,0 @@
-import type { Config } from "tailwindcss";
-
-const config: Config = {
-  content: [
-    "./src/pages/**/*.{js,ts,jsx,tsx,mdx}",
-    "./src/components/**/*.{js,ts,jsx,tsx,mdx}",
-    "./src/app/**/*.{js,ts,jsx,tsx,mdx}",
-  ],
-  theme: {
-    extend: {
-      colors: {
-        light: "var(--light)",
-        red: "var(--light-red)"
-      },
-    },
-  },
-  plugins: [],
-};
-export default config;
